# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Sistema de Notas, Arquitectura de Aplicaciones en la Nube",
        "version": "1.0.0"
    },
    "paths": {
        "/api/grades": {
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Registrar una nueva nota",
                "operationId": "efc1d4b5eaeb535c0441557a378fb4ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "subject_id",
                                    "grade"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "subject_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "grade": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nota registrada con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nota registrada con éxito"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "user_id": [
                                                    "El campo user_id es obligatorio."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/grades/{id}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Obtener las notas de un alumno",
                "operationId": "5d683dc34e7443ac2017ee6871375490",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del alumno",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de notas del alumno",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "subject_name": {
                                                "type": "string",
                                                "example": "Matemáticas"
                                            },
                                            "grade": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 8.5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "El alumno no tiene notas asignadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "El Alumno no tiene notas asignadas aún"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Grades"
                ],
                "summary": "Eliminar una nota",
                "operationId": "42569315e0d20107adf2fe7414fab888",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la nota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nota eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nota eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Nota no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "string",
                                            "example": "Nota no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Grades"
                ],
                "summary": "Actualizar una nota existente",
                "operationId": "44b8ac2c90bf993a13ce46b3c8e1f403",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la nota",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "grade"
                                ],
                                "properties": {
                                    "grade": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 9
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nota actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "string",
                                            "example": "Nota actualizada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Error de validación o nota no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "string",
                                            "example": "Nota no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/subjects": {
            "get": {
                "tags": [
                    "Subjects"
                ],
                "summary": "Get all subjects",
                "description": "Return a list of all subjects",
                "operationId": "c855f23389ed099cf68e0ec8b726d707",
                "responses": {
                    "200": {
                        "description": "List of subjects successfully obtained"
                    },
                    "204": {
                        "description": "No subjects registered"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Grades",
            "description": "API for grades"
        },
        {
            "name": "Subjects",
            "description": "API for subjects"
        }
    ]
}